Reintroduce a generic.py to define the generic dispatch methods.
Move the dispatch methods away from PathRegistry onto this base class.
Do this in a base class instead of in App directly so that we avoid
cyclic import dependencies.
