+++ morepath/path.py
@@ -7,7 +7,6 @@ See also :class:`morepath.directive.PathRegistry`
 
 
 from dectate import DirectiveError
-import reg
 from reg import arginfo
 try:
     # Python 2
@@ -16,6 +15,7 @@ except ImportError:
     from urllib.parse import urlencode, quote
 
 
+from .generic import GenericApp
 from .reify import reify
 from .cachingreg import RegRegistry
 from .traject import Path as TrajectPath, TrajectRegistry
@@ -60,65 +60,6 @@ class PathRegistry(TrajectRegistry):
         :return: a :class:`reg.Lookup` instance.
         """
         return self.reg_registry.caching_lookup
-
-    @reg.dispatch_method(reg.match_class('model', lambda model: model))
-    def _class_path(self, model, variables):
-        """Get the path for a model class.
-
-        :param model: model class or :class:`morepath.App` subclass.
-        :param variables: dictionary with variables to reconstruct
-        the path and URL paramaters from path pattern.
-        :return: a :class:`morepath.path.PathInfo` with path within this app,
-          or ``None`` if the path couldn't be determined.
-        """
-        return None
-
-    @reg.dispatch_method('obj')
-    def _path_variables(self, obj):
-        """Get variables to use in path generation.
-
-        :param obj: model object or :class:`morepath.App` instance.
-        :return: a dict with the variables to use for constructing the path,
-        or ``None`` if no such dict can be found.
-        """
-        return self._default_path_variables(obj)
-
-    @reg.dispatch_method('obj')
-    def _default_path_variables(self, obj):
-        """Get default variables to use in path generation.
-
-        Invoked if no specific ``path_variables`` is registered.
-
-        :param obj: model object for ::class:`morepath.App` instance.
-        :return: a dict with the variables to use for constructing the
-        path, or ``None`` if no such dict can be found.
-        """
-        return None
-
-    @reg.dispatch_method('obj')
-    def _deferred_link_app(self, mounted, obj):
-        """Get application used for link generation.
-
-        :param mounted: current :class:`morepath.App` instance.
-        :param obj: model object to link to.
-        :return: instance of :class:`morepath.App` subclass that handles
-        link generation for this model, or ``None`` if no app exists
-        that can construct link.
-        """
-        return None
-
-    @reg.dispatch_method(reg.match_class('model', lambda model: model))
-    def _deferred_class_link_app(self, mounted, model, variables):
-        """Get application used for link generation for a model class.
-
-        :param mounted: current :class:`morepath.App` instance.
-        :param model: model class
-        :param variables: dict of variables used to construct class link
-        :return: instance of :class:`morepath.App` subclass that handles
-        link generation for this model class, or ``None`` if no app exists
-        that can construct link.
-        """
-        return None
 
     def register_path(self, model, path,
                       variables, converters, required, get_converters,
@@ -210,7 +151,7 @@ class PathRegistry(TrajectRegistry):
         :param func: function that gets a model instance argument and
           returns a variables dict.
         """
-        self.reg_registry.register_function(PathRegistry._path_variables,
+        self.reg_registry.register_function(GenericApp._path_variables,
                                             func,
                                             obj=model)
 
@@ -228,13 +169,13 @@ class PathRegistry(TrajectRegistry):
         converters = converters or {}
         get_path = Path(path, factory_args, converters, absorb)
 
-        self.reg_registry.register_function(PathRegistry._class_path, get_path,
+        self.reg_registry.register_function(GenericApp._class_path, get_path,
                                             model=model)
 
         def default_path_variables(obj):
             return {name: getattr(obj, name) for name in factory_args}
         self.reg_registry.register_function(
-            PathRegistry._default_path_variables,
+            GenericApp._default_path_variables,
             default_path_variables,
             obj=model)
 
@@ -249,7 +190,7 @@ class PathRegistry(TrajectRegistry):
           does the link generation.
         """
         self.reg_registry.register_function(
-            PathRegistry._deferred_link_app, app_factory,
+            GenericApp._deferred_link_app, app_factory,
             obj=model)
 
     def register_defer_class_links(self, model, get_variables, app_factory):
@@ -265,7 +206,7 @@ class PathRegistry(TrajectRegistry):
         """
         self.register_path_variables(model, get_variables)
         self.reg_registry.register_function(
-            PathRegistry._deferred_class_link_app, app_factory,
+            GenericApp._deferred_class_link_app, app_factory,
             model=model)
 
 
@@ -365,7 +306,7 @@ class Path(object):
 
         :param model: model class. Not actually used in the
           implementation but used for dispatch in
-          :meth:`PathRegistry._class_path`.
+          :meth:`GenericApp._class_path`.
         :param variables: dict with the variables used in the path. each
           argument to the factory function should be represented.
         :return: :class:`PathInfo` instance representing the path.
