+++ morepath/tests/test_link.py
@@ -88,7 +88,7 @@ def test_class_path_without_variables(info):
     r.register_inverse_path(model=Foo,
                             path='/',
                             factory_args=set())
-    info = r._class_path(Foo, {})
+    info = app._class_path(Foo, {})
     assert info.path == ''
     assert info.parameters == {}
 
@@ -103,7 +103,7 @@ def test_class_path_with_variables(info):
     r.register_inverse_path(model=Foo,
                             path='/foos/{name}',
                             factory_args=set(['name']))
-    info = r._class_path(Foo, {'name': 'a'})
+    info = app._class_path(Foo, {'name': 'a'})
     assert info.path == 'foos/a'
     assert info.parameters == {}
 
@@ -118,7 +118,7 @@ def test_class_path_with_parameters(info):
     r.register_inverse_path(model=Foo,
                             path='/foos',
                             factory_args=set(['name']))
-    info = r._class_path(Foo, {'name': 'a'})
+    info = app._class_path(Foo, {'name': 'a'})
     assert info.path == 'foos'
     assert info.parameters == {'name': ['a']}
 
@@ -133,7 +133,7 @@ def test_class_path_variables_with_converters(info):
                             path='/foos/{value}',
                             factory_args=set(['value']),
                             converters={'value': Converter(int)})
-    info = r._class_path(Foo, {'value': 1})
+    info = app._class_path(Foo, {'value': 1})
     assert info.path == 'foos/1'
     assert info.parameters == {}
 
@@ -148,7 +148,7 @@ def test_class_path_parameters_with_converters(info):
                             path='/foos',
                             factory_args=set(['value']),
                             converters={'value': Converter(int)})
-    info = r._class_path(Foo, {'value': 1})
+    info = app._class_path(Foo, {'value': 1})
     assert info.path == 'foos'
     assert info.parameters == {'value': ['1']}
 
@@ -163,7 +163,7 @@ def test_class_path_absorb(info):
                             path='/foos',
                             factory_args=set(),
                             absorb=True)
-    info = r._class_path(Foo, {'absorb': 'bar'})
+    info = app._class_path(Foo, {'absorb': 'bar'})
     assert info.path == 'foos/bar'
     assert info.parameters == {}
 
@@ -177,8 +177,8 @@ def test_class_path_extra_parameters(info):
     r.register_inverse_path(model=Foo,
                             path='/foos',
                             factory_args=set())
-    info = r._class_path(Foo, {'extra_parameters': {'a': 'A',
-                                                    'b': 'B'}})
+    info = app._class_path(Foo, {'extra_parameters': {'a': 'A',
+                                                      'b': 'B'}})
     assert info.path == 'foos'
     assert info.parameters == {'a': ['A'], 'b': ['B']}
 
@@ -193,8 +193,8 @@ def test_class_path_extra_parameters_convert(info):
                             path='/foos',
                             factory_args=set(),
                             converters={'a': Converter(int)})
-    info = r._class_path(Foo,
-                         {'extra_parameters': {'a': 1,
-                                               'b': 'B'}})
+    info = app._class_path(Foo,
+                           {'extra_parameters': {'a': 1,
+                                                 'b': 'B'}})
     assert info.path == 'foos'
     assert info.parameters == {'a': ['1'], 'b': ['B']}
