+++ morepath/tests/test_querytool.py
@@ -67,7 +67,7 @@ def test_predicate_fallback():
     # there aren't any predicates for class_path
     r = objects(dectate.query_app(
         App, 'predicate_fallback',
-        dispatch='morepath.App.config.path_registry._class_path'))
+        dispatch='morepath.App._class_path'))
     assert r == []
 
     r = objects(dectate.query_app(App, 'predicate_fallback',
@@ -103,7 +103,7 @@ def test_predicate():
     # there aren't any predicates for class_path
     r = objects(dectate.query_app(
         App, 'predicate',
-        dispatch='morepath.App.config.path_registry._class_path'))
+        dispatch='morepath.App._class_path'))
     assert r == []
 
     r = objects(dectate.query_app(App, 'predicate',
